refactor: Update computeSample function to accept fileObject and kmerArgs
The computeSample function in worker.js has been updated to accept a fileObject parameter and a kmerArgs parameter. This allows for more flexibility in specifying the file, hash, filename, and kmer arguments when computing the sample. The fileObject is destructured to extract the necessary properties, while the kmerArgs are passed directly to the sketch function. This change improves the modularity and reusability of the computeSample function.

Refactor the computeSample function in worker.js to accept a fileObject parameter and a kmerArgs parameter.