Refactor: Update instructions for adding new species
The README.md file was updated to provide clearer instructions for adding new species to the project. The step to add a new database to [mrcdata](https://mrcdata.dide.ic.ac.uk/beebop) was removed, as it is now handled by the `download_db` script in *beebop_py*. The step to add a new species to `args.json` in *beebop_py* remains unchanged.